Factor out nodeset_selection module
