+++ crates/admin/src/cluster_controller/logs_controller.rs
@@ -8,7 +8,6 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use std::cmp::min;
 use std::collections::HashMap;
 use std::iter;
 use std::num::NonZeroU8;
@@ -17,7 +16,7 @@ use std::sync::Arc;
 use std::time::Duration;
 
 use rand::prelude::IteratorRandom;
-use rand::{thread_rng, Rng, RngCore};
+use rand::{thread_rng, RngCore};
 use tokio::sync::Semaphore;
 use tokio::task::JoinSet;
 use tracing::{debug, trace, trace_span, Instrument};
@@ -38,18 +37,20 @@ use restate_types::logs::metadata::{
 };
 use restate_types::logs::{LogId, Lsn, TailState};
 use restate_types::metadata_store::keys::BIFROST_CONFIG_KEY;
-use restate_types::nodes_config::NodesConfiguration;
 use restate_types::partition_table::PartitionTable;
 use restate_types::replicated_loglet::{
-    LocationScope, NodeSet, ReplicatedLogletId, ReplicatedLogletParams, ReplicationProperty,
+    NodeSet, ReplicatedLogletId, ReplicatedLogletParams, ReplicationProperty,
 };
 use restate_types::retries::{RetryIter, RetryPolicy};
 use restate_types::{logs, GenerationalNodeId, NodeId, PlainNodeId, Version, Versioned};
 
+use crate::cluster_controller::nodeset_selection::{
+    NodeSelectionError, NodeSetSelectionStrategy, NodeSetSelector,
+};
 use crate::cluster_controller::observed_cluster_state::ObservedClusterState;
 use crate::cluster_controller::scheduler;
 
-type Result<T, E = LogsControllerError> = std::result::Result<T, E>;
+pub type Result<T, E = LogsControllerError> = std::result::Result<T, E>;
 
 const FALLBACK_MAX_RETRY_DELAY: Duration = Duration::from_secs(5);
 
@@ -380,166 +381,6 @@ fn build_new_replicated_loglet_configuration(
             None
         }
     }
-}
-
-#[derive(
-    Debug, Clone, Eq, PartialEq, derive_more::Display, derive_more::Into, derive_more::IntoIterator,
-)]
-struct WritableNodeSet(NodeSet);
-
-impl WritableNodeSet {
-    /// Constructs a new nodeset consisting of only alive, read-write storage nodes.
-    pub fn from(cluster_state: &ObservedClusterState, nodes_config: &NodesConfiguration) -> Self {
-        Self(
-            cluster_state
-                .alive_nodes
-                .keys()
-                .copied()
-                .filter(|node_id| {
-                    nodes_config
-                        .get_log_server_storage_state(node_id)
-                        .can_write_to()
-                })
-                .collect(),
-        )
-    }
-
-    pub fn len(&self) -> usize {
-        self.0.len()
-    }
-}
-
-/// Nodeset selector for picking a set of storage nodes for a replicated loglet out of a
-/// broader pool of available nodes.
-///
-/// This selector can be reused once constructed to make multiple decision in a single scheduling iteration,
-/// if the node configuration and the replication settings are not changing.
-#[cfg(feature = "replicated-loglet")]
-#[derive(Clone)]
-struct NodeSetSelector<'a> {
-    nodes_config: &'a NodesConfiguration,
-    cluster_state: &'a ObservedClusterState,
-}
-
-impl<'a> NodeSetSelector<'a> {
-    fn new<'b: 'a>(
-        nodes_config: &'b NodesConfiguration,
-        cluster_state: &'b ObservedClusterState,
-    ) -> NodeSetSelector<'a> {
-        Self {
-            nodes_config,
-            cluster_state,
-        }
-    }
-
-    /// Picks a set of storage nodes for a replicated loglet out of the available pool. Only alive, writable
-    /// storage nodes will be used.
-    fn select<R: Rng + ?Sized>(
-        &self,
-        strategy: NodeSetSelectionStrategy,
-        replication_property: &ReplicationProperty,
-        rng: &mut R,
-        preferred_nodes: &NodeSet,
-    ) -> Result<NodeSet, NodeSelectionError> {
-        if replication_property.at_greatest_scope().0 != &LocationScope::Node {
-            // todo: add support for other location scopes
-            unimplemented!("only node-scoped replication is currently supported");
-        }
-
-        // Only consider alive, writable storage nodes.
-        let candidates = WritableNodeSet::from(self.cluster_state, self.nodes_config);
-
-        if candidates.len() < replication_property.num_copies().into() {
-            return Err(NodeSelectionError::InsufficientWriteableNodes);
-        }
-
-        // ReplicationFactor(f+1) implies a minimum of 2f+1 nodes. At this point we are only
-        // calculating the nodeset floor size, the actual size will be determined by the specific
-        // strategy in use.
-        assert!(
-            replication_property.num_copies() < u8::MAX >> 1,
-            "The replication factor implies a cluster size that exceeds the maximum supported size"
-        );
-        let min_fault_tolerant_nodeset_size =
-            (replication_property.num_copies() as usize - 1) * 2 + 1;
-        assert!(
-            min_fault_tolerant_nodeset_size >= replication_property.num_copies() as usize,
-            "The calculated minimum nodeset size can not be less than the replication factor"
-        );
-
-        let nodeset_target_size = match strategy {
-            NodeSetSelectionStrategy::InferredFaultToleranceStrict
-            | NodeSetSelectionStrategy::AllAvailable => min_fault_tolerant_nodeset_size,
-            NodeSetSelectionStrategy::CappedNodesetSize(size) => {
-                min(min_fault_tolerant_nodeset_size, size.get() as usize)
-            }
-        };
-
-        let nodeset = match strategy {
-            NodeSetSelectionStrategy::InferredFaultToleranceStrict
-            | NodeSetSelectionStrategy::CappedNodesetSize(_) => {
-                let mut nodes = preferred_nodes
-                    .iter()
-                    .copied()
-                    .filter(|node_id| candidates.0.contains(node_id))
-                    .choose_multiple(rng, nodeset_target_size);
-
-                if nodes.len() < nodeset_target_size {
-                    let remaining = nodeset_target_size - nodes.len();
-                    nodes.extend(
-                        candidates
-                            .into_iter()
-                            .filter(|node_id| !preferred_nodes.contains(node_id))
-                            .choose_multiple(rng, remaining),
-                    );
-                }
-
-                NodeSet::from_iter(nodes)
-            }
-
-            // We ignore preferred nodes under AllAvailable as we will select all healthy nodes anyway
-            NodeSetSelectionStrategy::AllAvailable => NodeSet::from_iter(candidates),
-        };
-
-        // todo: location-aware selection
-        if nodeset.len() < nodeset_target_size {
-            trace!(
-                selected_nodes_count = ?nodeset.len(),
-                ?nodeset_target_size,
-                "Failed to place replicated loglet: insufficient writeable nodes in the nodeset to meet availability goal"
-            );
-            return Err(NodeSelectionError::InsufficientWriteableNodes);
-        }
-
-        Ok(nodeset)
-    }
-}
-
-#[cfg(feature = "replicated-loglet")]
-#[derive(Debug, Clone, Default)]
-pub enum NodeSetSelectionStrategy {
-    /// Selects a nodeset size based on the replication factor. The nodeset size is 2f+1, working
-    /// backwards inferred from the assumption that replication factor is f+1. This flavor is strict
-    /// in that it will not configure a nodeset smaller than the inferred 2f+1 size, making it a
-    /// reasonable safe default.
-    #[default]
-    InferredFaultToleranceStrict,
-
-    /// Selects at most the specified set of nodes, while respecting the replication factor. To
-    /// tolerate failures, this size must be strictly greater than the replication factor. This
-    /// strategy allows for advanced operators to explicitly control both the lower bound (via
-    /// replication factor) and upper bound (via the nodeset cap) of loglet nodesets.
-    CappedNodesetSize(NonZeroU8),
-
-    /// Use all available storage nodes. This will lead to over-replication and excessive resource
-    /// utilisation in large clusters.
-    AllAvailable,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, thiserror::Error)]
-pub enum NodeSelectionError {
-    #[error("Insufficient writeable nodes in the nodeset")]
-    InsufficientWriteableNodes,
 }
 
 /// Representation of supported loglet configuration types.
@@ -1380,406 +1221,5 @@ impl<'a> scheduler::PartitionProcessorPlacementHints
                 | LogState::Sealed { .. }
                 | LogState::Provisioning { .. } => None,
             })
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::collections::HashSet;
-
-    use enumset::{enum_set, EnumSet};
-    use googletest::prelude::*;
-
-    use restate_types::nodes_config::{LogServerConfig, NodeConfig, Role, StorageState};
-    use restate_types::replicated_loglet::ReplicationProperty;
-    use restate_types::PlainNodeId;
-
-    use super::*;
-
-    #[test]
-    #[should_panic(
-        expected = "not implemented: only node-scoped replication is currently supported"
-    )]
-    fn test_select_log_servers_unsupported_replication_scope() {
-        let nodes: Vec<PlainNodeId> = vec![1.into()];
-        let replication =
-            ReplicationProperty::with_scope(LocationScope::Zone, 1.try_into().unwrap());
-
-        let mut nodes_config = NodesConfiguration::default();
-        nodes_config.upsert_node(node(
-            1,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::Provisioning,
-        ));
-
-        let observed_state = ObservedClusterState {
-            alive_nodes: nodes
-                .iter()
-                .copied()
-                .map(|id| (id, id.with_generation(1)))
-                .collect(),
-            ..Default::default()
-        };
-
-        let strategy = NodeSetSelectionStrategy::InferredFaultToleranceStrict;
-        let preferred_nodes = NodeSet::empty();
-        let rng = &mut thread_rng();
-        NodeSetSelector::new(&nodes_config, &observed_state)
-            .select(strategy, &replication, rng, &preferred_nodes)
-            .unwrap(); // panics
-    }
-
-    #[test]
-    fn test_select_log_servers_insufficient_capacity() {
-        let nodes: Vec<PlainNodeId> = vec![1.into(), 2.into(), 3.into()];
-        let replication =
-            ReplicationProperty::with_scope(LocationScope::Node, 2.try_into().unwrap());
-
-        let mut nodes_config = NodesConfiguration::default();
-        nodes_config.upsert_node(node(0, enum_set!(Role::Admin), StorageState::Disabled));
-        nodes_config.upsert_node(node(
-            1,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::Provisioning,
-        ));
-        nodes_config.upsert_node(node(
-            2,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::ReadWrite,
-        ));
-        nodes_config.upsert_node(node(
-            3,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::ReadOnly,
-        ));
-        nodes_config.upsert_node(node(4, enum_set!(Role::Worker), StorageState::Disabled));
-
-        let observed_state = ObservedClusterState {
-            partitions: Default::default(), // not needed here
-            alive_nodes: nodes
-                .iter()
-                .copied()
-                .map(|id| (id, id.with_generation(1)))
-                .collect(),
-            dead_nodes: HashSet::default(),
-            nodes_to_partitions: Default::default(),
-        };
-
-        let strategy = NodeSetSelectionStrategy::AllAvailable;
-        let preferred_nodes = NodeSet::empty();
-        let rng = &mut thread_rng();
-        let selection = NodeSetSelector::new(&nodes_config, &observed_state).select(
-            strategy,
-            &replication,
-            rng,
-            &preferred_nodes,
-        );
-
-        assert_eq!(
-            selection,
-            Err(NodeSelectionError::InsufficientWriteableNodes)
-        );
-    }
-
-    #[test]
-    fn test_select_log_servers_exceeds_min_replication_if_possible() {
-        let nodes: Vec<PlainNodeId> = vec![1.into(), 2.into(), 3.into()];
-        let replication =
-            ReplicationProperty::with_scope(LocationScope::Node, 2.try_into().unwrap());
-
-        let mut nodes_config = NodesConfiguration::default();
-        nodes_config.upsert_node(node(
-            1,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::ReadWrite,
-        ));
-        nodes_config.upsert_node(node(
-            2,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::ReadWrite,
-        ));
-        nodes_config.upsert_node(node(
-            3,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::ReadWrite,
-        ));
-
-        let observed_state = ObservedClusterState {
-            partitions: Default::default(), // not needed here
-            alive_nodes: nodes
-                .iter()
-                .copied()
-                .map(|id| (id, id.with_generation(1)))
-                .collect(),
-            dead_nodes: HashSet::default(),
-            nodes_to_partitions: Default::default(),
-        };
-
-        let strategy = NodeSetSelectionStrategy::AllAvailable;
-        let preferred_nodes = NodeSet::empty();
-        let rng = &mut thread_rng();
-        let selection = NodeSetSelector::new(&nodes_config, &observed_state).select(
-            strategy,
-            &replication,
-            rng,
-            &preferred_nodes,
-        );
-
-        assert!(selection.is_ok());
-        assert_eq!(selection.unwrap(), NodeSet::from_iter(nodes));
-    }
-
-    #[test]
-    fn test_select_log_servers_single_node_cluster() {
-        // Replicated loglets should work just fine in single-node clusters, with the FT strategy inferring that f=0
-        let nodes: Vec<PlainNodeId> = vec![1.into()];
-        let replication =
-            ReplicationProperty::with_scope(LocationScope::Node, 1.try_into().unwrap());
-
-        let mut nodes_config = NodesConfiguration::default();
-        nodes_config.upsert_node(node(
-            1,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::ReadWrite,
-        ));
-
-        let observed_state = ObservedClusterState {
-            partitions: Default::default(), // not needed here
-            alive_nodes: nodes
-                .iter()
-                .copied()
-                .map(|id| (id, id.with_generation(1)))
-                .collect(),
-            dead_nodes: HashSet::default(),
-            nodes_to_partitions: Default::default(),
-        };
-
-        let strategy = NodeSetSelectionStrategy::InferredFaultToleranceStrict;
-        let preferred_nodes = NodeSet::empty();
-        let rng = &mut thread_rng();
-        let selection = NodeSetSelector::new(&nodes_config, &observed_state).select(
-            strategy,
-            &replication,
-            rng,
-            &preferred_nodes,
-        );
-
-        assert!(selection.is_ok());
-        assert_eq!(selection.unwrap(), NodeSet::from_iter(nodes));
-    }
-
-    #[test]
-    fn test_select_log_servers_capped_nodeset_size() {
-        let nodes: Vec<PlainNodeId> = vec![1.into(), 2.into(), 3.into()];
-        let replication =
-            ReplicationProperty::with_scope(LocationScope::Node, 2.try_into().unwrap());
-
-        let mut nodes_config = NodesConfiguration::default();
-        nodes_config.upsert_node(node(
-            1,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::ReadWrite,
-        ));
-        nodes_config.upsert_node(node(
-            2,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::ReadWrite,
-        ));
-        nodes_config.upsert_node(node(
-            3,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::ReadWrite,
-        ));
-        nodes_config.upsert_node(node(
-            4,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::ReadWrite,
-        ));
-        nodes_config.upsert_node(node(
-            5,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::ReadWrite,
-        ));
-
-        let observed_state = ObservedClusterState {
-            partitions: Default::default(), // not needed here
-            alive_nodes: nodes
-                .iter()
-                .copied()
-                .map(|id| (id, id.with_generation(1)))
-                .collect(),
-            dead_nodes: HashSet::default(),
-            nodes_to_partitions: Default::default(),
-        };
-
-        let strategy = NodeSetSelectionStrategy::CappedNodesetSize(3.try_into().unwrap());
-        let preferred_nodes = NodeSet::empty();
-        let rng = &mut thread_rng();
-        let selection = NodeSetSelector::new(&nodes_config, &observed_state).select(
-            strategy,
-            &replication,
-            rng,
-            &preferred_nodes,
-        );
-
-        assert!(selection.is_ok());
-
-        let nodeset = selection.unwrap();
-        assert_eq!(nodeset.len(), 3);
-        nodeset.iter().for_each(|node_id| {
-            assert!(nodes.contains(node_id));
-        });
-    }
-
-    #[test]
-    fn test_select_log_servers_respects_preferred_nodes() {
-        let nodes: Vec<PlainNodeId> = vec![1.into(), 2.into(), 3.into(), 4.into(), 5.into()];
-        let replication =
-            ReplicationProperty::with_scope(LocationScope::Node, 2.try_into().unwrap());
-
-        let mut nodes_config = NodesConfiguration::default();
-        nodes_config.upsert_node(node(
-            1,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::DataLoss,
-        )); // this node is alive, but not writable
-        nodes_config.upsert_node(node(
-            2,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::ReadWrite,
-        )); // this node has stopped responding and is considered dead
-        nodes_config.upsert_node(node(
-            3,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::ReadWrite,
-        )); // this node was previously used
-        for n in 4..=10 {
-            nodes_config.upsert_node(node(
-                n,
-                enum_set!(Role::LogServer | Role::Worker),
-                StorageState::ReadWrite,
-            ));
-        }
-
-        let observed_state = ObservedClusterState {
-            partitions: Default::default(), // not needed here
-            alive_nodes: nodes
-                .iter()
-                .copied()
-                .filter(|id| *id != PlainNodeId::new(2))
-                .map(|id| (id, id.with_generation(1)))
-                .collect(),
-            dead_nodes: HashSet::<PlainNodeId, Xxh3Builder>::from_iter([PlainNodeId::new(2)]),
-            nodes_to_partitions: Default::default(),
-        };
-
-        let strategy = NodeSetSelectionStrategy::InferredFaultToleranceStrict;
-        let preferred_nodes =
-            NodeSet::from_iter::<Vec<PlainNodeId>>(vec![1.into(), 2.into(), 3.into()]);
-        let rng = &mut thread_rng();
-        let selection = NodeSetSelector::new(&nodes_config, &observed_state).select(
-            strategy,
-            &replication,
-            rng,
-            &preferred_nodes,
-        );
-
-        println!("{:?}", selection);
-        assert!(selection.is_ok());
-
-        let nodeset = selection.unwrap();
-        assert_that!(nodeset, contains(eq(PlainNodeId::new(3))));
-        assert_that!(nodeset, not(contains(eq(PlainNodeId::new(1)))));
-        assert_that!(nodeset, not(contains(eq(PlainNodeId::new(2)))));
-    }
-
-    /// In this test we have a cluster with 4 nodes, of which three were previously used for a loglet.
-    /// The scenario is to verify that the nodeset selector avoids scheduling a loglet on the only two viable
-    /// log servers. Although this meets the basic replication requirement, it would render the loglet vulnerable
-    /// to getting stuck if a single node from its nodeset fails.
-    #[test]
-    fn test_select_log_servers_avoids_degraded_nodes() {
-        let nodes: Vec<PlainNodeId> = vec![1.into(), 2.into(), 3.into(), 4.into()];
-        let replication =
-            ReplicationProperty::with_scope(LocationScope::Node, 2.try_into().unwrap());
-
-        let mut nodes_config = NodesConfiguration::default();
-        nodes_config.upsert_node(node(
-            1,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::DataLoss,
-        )); // this node is alive, but not writable
-        nodes_config.upsert_node(node(
-            2,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::ReadWrite,
-        )); // this node is in the correct storage state but has stopped responding and is considered dead (see below)
-        nodes_config.upsert_node(node(
-            3,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::ReadWrite,
-        )); // this node was previously used
-        nodes_config.upsert_node(node(
-            4,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::ReadWrite,
-        )); // new node
-
-        let mut observed_state = ObservedClusterState {
-            partitions: Default::default(), // not needed here
-            alive_nodes: nodes
-                .iter()
-                .copied()
-                .filter(|id| *id != PlainNodeId::new(2))
-                .map(|id| (id, id.with_generation(1)))
-                .collect(),
-            dead_nodes: HashSet::<PlainNodeId, Xxh3Builder>::from_iter([PlainNodeId::new(2)]),
-            nodes_to_partitions: Default::default(),
-        };
-
-        let preferred_nodes =
-            NodeSet::from_iter::<Vec<PlainNodeId>>(vec![1.into(), 2.into(), 3.into()]);
-        let rng = &mut thread_rng();
-
-        let selection = NodeSetSelector::new(&nodes_config, &observed_state).select(
-            NodeSetSelectionStrategy::InferredFaultToleranceStrict,
-            &replication,
-            rng,
-            &preferred_nodes,
-        );
-        assert!(selection.is_err());
-
-        // one additional healthy node should make it possible to find a viable nodeset
-        nodes_config.upsert_node(node(
-            5,
-            enum_set!(Role::LogServer | Role::Worker),
-            StorageState::ReadWrite,
-        ));
-        observed_state
-            .alive_nodes
-            .insert(5.into(), PlainNodeId::new(5).with_generation(1));
-        let selection = NodeSetSelector::new(&nodes_config, &observed_state).select(
-            NodeSetSelectionStrategy::InferredFaultToleranceStrict,
-            &replication,
-            rng,
-            &preferred_nodes,
-        );
-        assert!(selection.is_ok());
-        assert_eq!(selection.unwrap().len(), 3);
-    }
-
-    pub fn node(
-        id: impl Into<PlainNodeId>,
-        roles: EnumSet<Role>,
-        storage_state: StorageState,
-    ) -> NodeConfig {
-        let id: PlainNodeId = id.into();
-        NodeConfig::new(
-            format!("n{}", id),
-            id.with_generation(1),
-            format!("http://n{}", id).parse().unwrap(),
-            roles,
-            LogServerConfig { storage_state },
-        )
     }
 }
